% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{ml_init_context}
\alias{ml_init_context}
\alias{ml_ingest}
\title{Internal functions}
\usage{
ml_init_context(x, env = env_parent(caller_env()), stack = env_get(env,
  "MLSTACK.."), call = caller_call(), fn = caller_fn())

ml_ingest(x = identity, ...)
}
\arguments{
\item{x}{ml context}

\item{env}{env containing MLSTACK..}

\item{stack}{MLSTACK..}

\item{call}{caller call}

\item{fn}{caller fn}
}
\description{
These functions are exported due to the nature of the `multiline` dispatch.
}
\keyword{internal}
