% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magistral.R
\docType{package}
\name{magistral-package}
\alias{magistral}
\alias{magistral-package}
\title{Lightweight machine learning and data processing pipeline}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Vitalie Spinu \email{spinuvit@gmail.com}

}
\keyword{internal}
