% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vars.R
\name{refval}
\alias{refval}
\alias{refval_extractor}
\title{Extract reference value from a vector}
\usage{
refval(x, type = c("median", "mean", "min", "min-", "max", "max+"))

refval_extractor(type = c("median", "mean", "min", "min-", "max", "max+"))
}
\description{
Extract reference value from a vector
}
